requirements.txt
Flask==2.0.1
requests==2.26.0
pandas==1.3.3
python-dotenv==0.19.0
numpy==1.21.2